requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
sqlalchemy==2.0.30
pydantic==2.7.1
pydantic-settings==2.2.1
alembic==1.13.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
